Update ca_rewrite to use hook_node_presave instead of hook_node_view_alter
For each node that is loaded within a view, the ca_rewrite_node_view_alter() must parse and filter the node title on-the-fly, which is somewhat expensive on the homepage where there are many nodes being loaded. This on-the-fly processing/filtering of node titles involves the following:
1. For each node in the view, query the database to see what pattern rewrites are available
2. Iterate over each pattern and filter the node title with each rewrite pattern before rendering to page
3. Title never changes in the DB
4. Cache is never fully leveraged due to the dynamic nature of the filtering.

This experiment attempts to improve performance/memory by only changing the node title once -- when the node is imported from the feed. View results and rendering are then able to be better cached.

Because this approach only affects new nodes being imported, I've also added a hook_update_N to the module install file to batch update any existing feed_item nodes with the ca_rewrite filtering.